+++ src/Rule/Trait/CountableLimitTrait.php
@@ -130,14 +130,18 @@ trait CountableLimitTrait
             throw new InvalidArgumentException('Only positive or zero values are allowed.');
         }
 
+        $messageForUsingExactly = 'Use $exactly instead.';
+
         if ($this->min !== null && $this->max !== null) {
             if ($this->min > $this->max) {
                 throw new InvalidArgumentException('$min must be lower than $max.');
             }
 
             if ($this->min === $this->max) {
-                throw new InvalidArgumentException('Use $exactly instead.');
+                throw new InvalidArgumentException($messageForUsingExactly);
             }
+        } elseif ($this->min === null && $this->max === 0) {
+            throw new InvalidArgumentException($messageForUsingExactly);
         }
     }