+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Bug #586: Allow `0` as a limit in `CountableLimitTrait` (@arogachev)
+- Bug #431: Add support for empty iterables to `WhenEmpty` (@arogachev)
 
 ## 1.0.0 February 22, 2023