+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Validator Change Log
 
-## 1.0.1 under development
+## 1.1.0 under development
 
 - Enh #567: Add immutable setter for modifying `Validator::$defaultSkipOnEmptyCondition` (@arogachev)
 - Bug #586: Allow `0` as a limit in `CountableLimitTrait` (@arogachev)